+++ output.tf
@@ -5,3 +5,15 @@ output "connect_command" {
 output "enable_keyvault_addon_command" {
   value = "az aks enable-addons --addons azure-keyvault-secrets-provider --name ${azurerm_kubernetes_cluster.aks.name} --resource-group ${azurerm_resource_group.public.name}"
 }
+
+output "key_vault_id" {
+  value = azurerm_key_vault.public.id
+}
+
+output "node_resource_group" {
+  value = azurerm_kubernetes_cluster.aks.node_resource_group
+}
+
+output "key_vault_uami_client_id" {
+  value = data.azurerm_user_assigned_identity.kv_uami.client_id
+}