+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Enable keyvault addon in AKS
 - Configure csi provider class for keyvault
 - Create a pod with mounts pointing to secrets
+- Add local exec provisioner for AKS keyvault addon